refactor: adjust naming
