+++ src/store/rootReducer.ts
@@ -1,5 +1,5 @@
 import { channelsSlice } from "src/features/channels/channelsSlice";
-import { requestedPackagesSlice } from "src/features/requestedPackages/requestedPackages";
+import { requestedPackagesSlice } from "src/features/requestedPackages/requestedPackagesSlice";
 
 export const rootReducer = {
   channels: channelsSlice.reducer,
